cargo + clippy

+++ jolt-core/src/zkvm/r1cs/ops.rs
@@ -96,9 +96,9 @@ impl LC {
             .iter()
             .map(|term| match term.0 {
                 Variable::Input(var_index) => {
-                    term.1 as i128 * flattened_polynomials[var_index].get_coeff_i128(row)
+                    term.1 * flattened_polynomials[var_index].get_coeff_i128(row)
                 }
-                Variable::Constant => term.1 as i128,
+                Variable::Constant => term.1,
             })
             .sum()
     }
